% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{componentItems}
\alias{componentItems}
\title{Get items for a speaker from a component}
\usage{
componentItems(speakers, component)
}
\arguments{
\item{speakers}{A vector of speaker identifiers, eg. "1_10"}

\item{component}{The Austalk component name, eg. 'digits', see \code{components()} for a full list}
}
\value{
a dataframe containing item URLs, speaker id and item id
}
\description{
Get items for a speaker from a component
}
