% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{downloadItems}
\alias{downloadItems}
\title{Download documents for a list of items}
\usage{
downloadItems(items, destination, channels = c("speaker16"))
}
\arguments{
\item{items}{A dataframe containing at least an \code{item_url} column}

\item{destination}{Directory to save downloaded files}

\item{channels}{A vector of channel names or patterns matching files to be downloaded}
}
\value{
Files are downloaded
}
\description{
Download documents for a list of items
}
