% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{getContribution}
\alias{getContribution}
\title{Get list of items from a contribution}
\usage{
getContribution(url)
}
\arguments{
\item{url}{The URL of the contribution}
}
\value{
A data frame containing item urls
}
\description{
Get list of items from a contribution
}
