% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{getItemList}
\alias{getItemList}
\title{Get a list of items from an item list}
\usage{
getItemList(url)
}
\arguments{
\item{url}{The URL of the item list}
}
\value{
A data frame containing the item URLs
}
\description{
Get a list of items from an item list
}
