% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{prompts}
\alias{prompts}
\title{Return the prompts for a given component}
\usage{
prompts(component)
}
\arguments{
\item{component}{The name of the target component e.g. digits}
}
\value{
A data frame containing the prompts and item numbers for this component
}
\description{
Return the prompts for a given component
}
