% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakers.R
\name{speakerItems}
\alias{speakerItems}
\title{Generate a table of the number of recordings in components for
all speakers}
\usage{
speakerItems(components)
}
\arguments{
\item{components}{a vector of component names, eg. c('digits', 'words-1'), see \code{components()} for a full list}
}
\value{
a data.frame containing speaker ids and a count of items for each component
}
\description{
Generate a table of the number of recordings in components for
all speakers
}
